requirements.txt
aiogram==3.13.1
python-dotenv==1.0.0
loguru==0.7.2               # опционально, для красивого лога